Enhance the metadata caching to store all the required reflection info within static field.
