+++ src/test/java/ru/yandex/qatools/fsm/FallenRaisedStateMachineTest.java
@@ -16,7 +16,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
 /**
- * @author: Ilya Sadykov (mailto: smecsia@yandex-team.ru)
+ * @author: Ilya Sadykov
  */
 
 public class FallenRaisedStateMachineTest {
