+++ src/test/java/ru/yandex/qatools/fsm/utils/ReflectUtilsTest.java
@@ -12,7 +12,7 @@ import static ru.yandex.qatools.fsm.utils.ReflectUtils.collectAllSuperclassesAnd
 import static ru.yandex.qatools.fsm.utils.ReflectUtils.containsClass;
 
 /**
- * @author: Ilya Sadykov (mailto: smecsia@yandex-team.ru)
+ * @author: Ilya Sadykov
  */
 public class ReflectUtilsTest {
